feature: 专题模块框架及确定实体类定义

+++ service-topic/src/main/java/com/dataint/topic/model/vo/CrawlSiteVO.java
@@ -6,7 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CrawlSiteVO {
-    private Integer siteId;
+
+    private Long siteId;
 
     private String siteName;
